+++ src/main/java/com/example/Catalog/Managment/Repository/CustomerRepository.java
@@ -3,7 +3,7 @@ package com.example.Catalog.Managment.Repository;
 import com.example.Catalog.Managment.Entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer,Integer>
+public interface CustomerRepository extends JpaRepository<Customer,Long>
 {
 
 }
