+++ src/main/java/com/example/Catalog/Managment/Repository/OrderItemRepository.java
@@ -3,5 +3,5 @@ package com.example.Catalog.Managment.Repository;
 import com.example.Catalog.Managment.Entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 }
