+++ src/main/java/com/example/Catalog/Managment/Service/CustomerService.java
@@ -9,6 +9,6 @@ import java.util.List;
 public interface CustomerService
 {
     ResponseEntity<ApiResponse<CustomerDto>> create(CustomerDto customerDto);
-    ResponseEntity<ApiResponse<CustomerDto>> getById(int id);
+    ResponseEntity<ApiResponse<CustomerDto>> getById(Long id);
     ResponseEntity<ApiResponse<List<CustomerDto>>> getAll();
 }
